Object for managing backing up orderly db

Object for managing backing up orderly db

Object for managing backing up orderly db

Public fields

config

Orderly config

backup_period

How frequently should backup be run

last_backup

Date-time when last backup was run

Methods

Public methods

orderly_backup_$new()

orderly_backup_$check_backup()

orderly_backup_$do_backup()

Method new()

Create object, and do first backup.

Usage

orderly_backup_$new(config, backup_period)

Arguments

config

Orderly config.

backup_period

Period (in seconds) between DB backups. This is a guide only as backups happen on a API preroute - if more than this many seconds have elapsed when the API receives a request then a backup of the db will be performed. This creates a copy of orderly's destination database inbackup/db with the same filename as the destination database, even if that database typically lives outside of the orderly tree. In case of corruption of the database, this backup can be manually moved into place. This is only needed if you are storing information alongside the core orderly tables (as done by OrderlyWeb). If NULL then backup is skipped.

Method check_backup()

Check if a backup is due to be run, and run it if it is due.

Usage

orderly_backup_$check_backup()

Method do_backup()

Backup the orderly database.

Usage

orderly_backup_$do_backup()