Orderly runner

An orderly runner. This is used to run reports as a server process. It's designed to be used in conjunction with OrderlyWeb, so there is no "draft" stage and reports are committed as soon as they are run. This function is not intended for human end users, only for creating automated tools for use with orderly.

orderly_runner(path, allow_ref = NULL, queue_id = NULL, workers = 1)

Arguments

path
Path to use

allow_ref
Allow git to change branches/ref for run. If not given, then we will look to see if the orderly configuration disallows branch changes (based on the ORDERLY_API_SERVER_IDENTITY environment variable and the master_only setting of the relevant server block.

queue_id
ID of an existing queue to connect to, creates a new one if NULL

workers
Number of workers to spawn

Value

A runner object, with methods designed for internal use only.

Examples

available <- redux::redis_available()
if (available) {
  path <- orderly:::prepare_orderly_git_example()
  runner <- orderly.server::orderly_runner(path[["local"]], workers = 0)
}

#> Writing to '/tmp/RtmpZ9JyE9/file1809a4ad140c4/demo/.gitignore'

#> [ git        ]  checkout other; was master

#> [ git        ]  checkout master; was other

#> [ git        ]  checkout master; was master

#> [ git        ]  checkout master; was master

#> [ git        ]  fetch

#> [ git        ]  clone

#> Connecting to redis at redis://127.0.0.1:6379

#> Starting queue